LayoutTests/http/tests/appcache/manifest-redirect-expected.txt

CONSOLE MESSAGE: line 0: Application Cache manifest could not be fetched, because a redirection was attempted.
Test that a redirect makes resource caching fail.

Should say SUCCESS:

SUCCESS
